+++ test/openmaize/confirm_email_test.exs
@@ -3,7 +3,7 @@ defmodule Openmaize.ConfirmEmailTest do
   use Plug.Test
 
   import Ecto.Changeset
-  alias Openmaize.{ConfirmEmail, TestRepo, TestUser}
+  alias Openmaize.{ConfirmEmail, EctoDB, TestRepo, TestUser}
 
   @valid_link "email=fred%2B1%40mail.com&key=lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw"
   @name_link "username=fred&key=lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw"
@@ -29,37 +29,37 @@ defmodule Openmaize.ConfirmEmailTest do
   end
 
   test "confirmation succeeds for valid token" do
-    conn = call_confirm(@valid_link, {120, :email, nil})
+    conn = call_confirm(@valid_link, {EctoDB, 120, :email, nil})
     assert user_confirmed()
     assert conn.private.openmaize_info =~ "Account successfully confirmed"
   end
 
   test "confirmation fails for invalid token" do
-    conn = call_confirm(@invalid_link, {120, :email, nil})
+    conn = call_confirm(@invalid_link, {EctoDB, 120, :email, nil})
     refute user_confirmed()
     assert conn.private.openmaize_error =~ "Confirmation for"
   end
 
   test "confirmation fails for expired token" do
-    conn = call_confirm(@valid_link, {0, :email, nil})
+    conn = call_confirm(@valid_link, {EctoDB, 0, :email, nil})
     refute user_confirmed()
     assert conn.private.openmaize_error =~ "Confirmation for"
   end
 
   test "invalid link error" do
-    conn = call_confirm(@incomplete_link, {120, :email, nil})
+    conn = call_confirm(@incomplete_link, {EctoDB, 120, :email, nil})
     refute user_confirmed()
     assert conn.private.openmaize_error =~ "Invalid link"
   end
 
   test "confirmation succeeds with different unique id" do
-    conn = call_confirm(@name_link, {120, :username, nil})
+    conn = call_confirm(@name_link, {EctoDB, 120, :username, nil})
     assert user_confirmed()
     assert conn.private.openmaize_info =~ "Account successfully confirmed"
   end
 
   test "confirmation fails when query fails" do
-    conn = call_confirm("key=lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw", {120, :email, nil})
+    conn = call_confirm("key=lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw", {EctoDB, 120, :email, nil})
     refute user_confirmed()
     assert conn.private.openmaize_error =~ "Confirmation for"
   end
@@ -74,6 +74,12 @@ defmodule Openmaize.ConfirmEmailTest do
     {key, link} = ConfirmEmail.gen_token_link("fred", :username)
     assert link =~ "username=fred&key="
     assert :binary.match(link, [key]) == {18, 32}
+  end
+
+  test "raises error if no db_module is set" do
+    assert_raise ArgumentError, "You need to set the db_module value for Openmaize.ConfirmEmail", fn ->
+      call_confirm(@valid_link, {nil, 120, :email, nil})
+    end
   end
 
 end