+++ test/openmaize/login_test.exs
@@ -2,7 +2,7 @@ defmodule Openmaize.LoginTest do
   use ExUnit.Case
   use Plug.Test
 
-  alias Openmaize.{DummyCrypto, Login, Login.Name}
+  alias Openmaize.{DummyCrypto, EctoDB, Login, Login.Name}
   alias OpenmaizeJWT.{Tools, Verify}
 
   def call(name, password, uniq, opts) do
@@ -12,7 +12,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "login succeeds with username" do
-    opts = {:cookie, :username, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, :username, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("ray", "h4rd2gU3$$", "username", opts)
     assert conn.resp_cookies["access_token"]
     refute conn.private[:openmaize_error]
@@ -20,7 +20,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "login succeeds with email" do
-    opts = {:cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("ray@mail.com", "h4rd2gU3$$", "email", opts)
     assert conn.resp_cookies["access_token"]
     refute conn.private[:openmaize_error]
@@ -29,7 +29,7 @@ defmodule Openmaize.LoginTest do
 
   test "login fails when crypto mod changes" do
     Application.put_env(:openmaize, :crypto_mod, DummyCrypto)
-    opts = {:cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("ray@mail.com", "h4rd2gU3$$", "email", opts)
     refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
@@ -39,7 +39,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "login fails for incorrect password" do
-    opts = {:cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("ray@mail.com", "oohwhatwasitagain", "email", opts)
     refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
@@ -47,7 +47,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "login fails for invalid email" do
-    opts = {:cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("dick@mail.com", "h4rd2gU3$$", "email", opts)
     refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
@@ -55,7 +55,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "multiple possible unique ids - email for email_username func" do
-    opts = {:cookie, &Name.email_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, &Name.email_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("ray@mail.com", "h4rd2gU3$$", "email", opts)
     assert conn.resp_cookies["access_token"]
     refute conn.private[:openmaize_error]
@@ -63,7 +63,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "multiple possible unique ids - username for email_username func" do
-    opts = {:cookie, &Name.email_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, &Name.email_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("ray", "h4rd2gU3$$", "email", opts)
     assert conn.resp_cookies["access_token"]
     refute conn.private[:openmaize_error]
@@ -71,7 +71,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "multiple possible unique ids - phone for phone_username func" do
-    opts = {:cookie, &Name.phone_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, &Name.phone_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("081555555", "h4rd2gU3$$", "phone", opts)
     assert conn.resp_cookies["access_token"]
     refute conn.private[:openmaize_error]
@@ -79,7 +79,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "multiple possible unique ids - username for phone_username func" do
-    opts = {:cookie, &Name.phone_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, &Name.phone_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("ray", "h4rd2gU3$$", "phone", opts)
     assert conn.resp_cookies["access_token"]
     refute conn.private[:openmaize_error]
@@ -87,7 +87,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "fail login with multiple possible unique ids - phone for phone_username func" do
-    opts = {:cookie, &Name.phone_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, &Name.phone_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("081555555", "oohwhatwasitagain", "phone", opts)
     refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
@@ -95,7 +95,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "fail login with multiple possible unique ids - username for phone_username func" do
-    opts = {:cookie, &Name.phone_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
+    opts = {EctoDB, :cookie, &Name.phone_username/1, &OpenmaizeJWT.Plug.add_token/5, nil}
     conn = call("rav", "h4rd2gU3$$", "phone", opts)
     refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
@@ -103,7 +103,7 @@ defmodule Openmaize.LoginTest do
   end
 
   test "override default token validity, remember me example" do
-    opts = {:cookie, :username, &OpenmaizeJWT.Plug.add_token/5, 10_080}
+    opts = {EctoDB, :cookie, :username, &OpenmaizeJWT.Plug.add_token/5, 10_080}
     conn = conn(:post, "/login", %{"user" => %{"username" => "ray",
        "password" => "h4rd2gU3$$", "remember_me" => "true"}}) |> Login.call(opts)
     token = conn.resp_cookies["access_token"]
@@ -113,13 +113,20 @@ defmodule Openmaize.LoginTest do
   end
 
   test "remember me set to false example" do
-    opts = {:cookie, :username, &OpenmaizeJWT.Plug.add_token/5, 10_080}
+    opts = {EctoDB, :cookie, :username, &OpenmaizeJWT.Plug.add_token/5, 10_080}
     conn = conn(:post, "/login", %{"user" => %{"username" => "ray",
        "password" => "h4rd2gU3$$", "remember_me" => "false"}}) |> Login.call(opts)
     token = conn.resp_cookies["access_token"]
     assert token.max_age == 7200
     {:ok, %{exp: exp}} = Verify.verify_token token.value
     assert exp - Tools.current_time < 8_000_000
+  end
+
+  test "raises error if no db_module is set" do
+    opts = {nil, :cookie, :email, &OpenmaizeJWT.Plug.add_token/5, nil}
+    assert_raise ArgumentError, "You need to set the db_module value for Openmaize.Login", fn ->
+      call("ray@mail.com", "h4rd2gU3$$", "email", opts)
+    end
   end
 
 end