+++ test/openmaize/onetime_pass_test.exs
@@ -3,7 +3,7 @@ defmodule Openmaize.OnetimePassTest do
   use Plug.Test
 
   alias Comeonin.Otp
-  alias Openmaize.OnetimePass
+  alias Openmaize.{EctoDB, OnetimePass}
   alias OpenmaizeJWT.{Tools, Verify}
 
   def get_count do
@@ -17,26 +17,26 @@ defmodule Openmaize.OnetimePassTest do
 
   test "check hotp with default options" do
     user = %{"hotp" => "816065", "storage" => "cookie", "uniq" => "username", "id" => "5"}
-    conn = call(user, {&OpenmaizeJWT.Plug.add_token/5, []})
+    conn = call(user, {EctoDB, &OpenmaizeJWT.Plug.add_token/5, []})
     assert conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_info] == 2
     refute conn.private[:openmaize_error]
     fail = %{"hotp" => "816066", "storage" => "cookie", "uniq" => "username",
      "id" => "5", "override_exp" => nil}
-    conn = call(fail, {&OpenmaizeJWT.Plug.add_token/5, []})
+    conn = call(fail, {EctoDB, &OpenmaizeJWT.Plug.add_token/5, []})
     refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
   end
 
   test "check hotp with last option" do
     user = %{"hotp" => "088239", "storage" => "cookie", "uniq" => "username", "id" => "5"}
-    conn = call(user, {&OpenmaizeJWT.Plug.add_token/5, [last: 18]})
+    conn = call(user, {EctoDB, &OpenmaizeJWT.Plug.add_token/5, [last: 18]})
     assert conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_info] == 19
     refute conn.private[:openmaize_error]
     fail = %{"hotp" => "088238", "storage" => "cookie", "uniq" => "username",
      "id" => "5", "override_exp" => nil}
-    conn = call(fail, {&OpenmaizeJWT.Plug.add_token/5, [last: 18]})
+    conn = call(fail, {EctoDB, &OpenmaizeJWT.Plug.add_token/5, [last: 18]})
     refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
   end
@@ -45,7 +45,7 @@ defmodule Openmaize.OnetimePassTest do
     token = Otp.gen_totp("MFRGGZDFMZTWQ2LK")
     user = %{"totp" => token, "storage" => "cookie", "uniq" => "email",
      "id" => "5", "override_exp" => nil}
-    conn = call(user, {&OpenmaizeJWT.Plug.add_token/5, []})
+    conn = call(user, {EctoDB, &OpenmaizeJWT.Plug.add_token/5, []})
     assert conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_info]
     refute conn.private[:openmaize_error]
@@ -55,11 +55,18 @@ defmodule Openmaize.OnetimePassTest do
     token = Otp.gen_totp("MFRGGZDFMZTWQ2LK")
     user = %{"totp" => token, "storage" => "cookie", "uniq" => "email",
      "id" => "5", "override_exp" => "10080"}
-    conn = call(user, {&OpenmaizeJWT.Plug.add_token/5, []})
+    conn = call(user, {EctoDB, &OpenmaizeJWT.Plug.add_token/5, []})
     token = conn.resp_cookies["access_token"]
     assert token.max_age == 604_800
     {:ok, %{exp: exp}} = Verify.verify_token token.value
     assert exp - Tools.current_time > 500_000_000
+  end
+
+  test "raises error if no db_module is set" do
+    user = %{"hotp" => "816065", "storage" => "cookie", "uniq" => "username", "id" => "5"}
+    assert_raise ArgumentError, "You need to set the db_module value for Openmaize.OnetimePass", fn ->
+      call(user, {nil, &OpenmaizeJWT.Plug.add_token/5, []})
+    end
   end
 
 end